countNA {vtable}	R Documentation

Number of missing values in a vector

Description

This function calculates the number of values in a vector that are NA.

Usage

countNA(x)

Arguments

x
A vector.

Details

This function just shorthand for sum(is.na(x)), with a shorter name for reference in the vtable or sumtable summ option.

Examples

x <- c(1, 1, NA, 2, 3, NA)
countNA(x)

[Package vtable version 1.4.8 Index]