nuniq {vtable}	R Documentation

Number of unique values in a vector

Description

This function takes a vector and returns the number of unique values in that vector.

Usage

nuniq(x)

Arguments

x
A vector.

Details

This function is just shorthand for length(unique(x)), with a shorter name for reference in the vtable or sumtable summ option.

Examples

x <- c(1, 1, 2, 3, 4, 4, 4)
nuniq(x)

[Package vtable version 1.4.8 Index]