pctile {vtable}	R Documentation

Returns a vector of 100 percentiles

Description

This function calculates 100 percentiles of a vector and returns all of them.

Usage

pctile(x)

Arguments

x
A vector.

Details

This function just shorthand for quantile(x,1:100/100), with a shorter name for reference in the vtable or sumtable summ option, and which works with sumtable summ.names styling.

Examples

x <- 1:500
pctile(x)[50]
quantile(x,.5)
median(x)

[Package vtable version 1.4.8 Index]