propNA {vtable}	R Documentation

Proportion or number of missing values in a vector

Description

This function calculates the proportion of values in a vector that are NA.

Usage

propNA(x)

Arguments

x
A vector.

Details

This function just shorthand for mean(is.na(x)), with a shorter name for reference in the vtable or sumtable summ option.

Examples

x <- c(1, 1, NA, 2, 3, NA)
propNA(x)

[Package vtable version 1.4.8 Index]